+++ doc/sphinx/src/news.rst
@@ -32,6 +32,10 @@ Deprecated
 
 New features
 ^^^^^^^^^^^^
+* Execution
+
+  * A task parameter can be a ``std::vector`` of a FleCSI type such as a field accessor or a ``std::tuple`` that includes such a type (recursively).
+
 * Topologies
 
   * ``narray`` specializations need not define ``dimension``.